Migliorata suddivisione intefaccia-logica.
Rimosse proprietà ListaPedine e ListaBanconote dalla classe Tabellone.
Rivisto passaggio parametri dalle varie schermate MainWindow-FinestraDiGioco.
Rivisti alcuni costruttori.
Ora la classe Gioco sa dell'esistenza della classe Server ma non viceversa.
Da sistemare metodi nella classe Server.

+++ ProgettoMonopoly/ProgettoMonopoly/Classi/Contratti/ContrattoImpresaPubblica.cs
@@ -7,7 +7,7 @@ namespace ProgettoMonopoly
 {
     public class ContrattoImpresaPubblica : Contratto
     {
-        public ContrattoImpresaPubblica(string nomeContratto, float valoreContratto, List<int> rendita) : base(nomeContratto, valoreContratto, rendita)
+        public ContrattoImpresaPubblica(string nomeContratto, int valoreContratto, List<int> rendita) : base(nomeContratto, valoreContratto, rendita)
         {
 
         }
